[Refactoring] Move Model.Simulator in the root src/netzob/ directory
