+++ src/netzob/Simulator/Channels/AbstractChannel.py
