+++ src/netzob/Simulator/Channels/UDPServer.py
@@ -35,7 +35,6 @@
 #| Standard library imports                                                  |
 #+---------------------------------------------------------------------------+
 import socket
-import time
 
 #+---------------------------------------------------------------------------+
 #| Related third party imports                                               |
@@ -45,20 +44,23 @@ import time
 #| Local application imports                                                 |
 #+---------------------------------------------------------------------------+
 from netzob.Common.Utils.Decorators import typeCheck, NetzobLogger
-from netzob.Model.Simulator.Channels.AbstractChannel import AbstractChannel, ChannelDownException
+from netzob.Simulator.Channels.AbstractChannel import AbstractChannel
 
 
 @NetzobLogger
-class TCPClient(AbstractChannel):
-    """A TCPClient is a communication channel. It allows to create client connecting
-    to a specific IP:Port server over a TCP socket.
+class UDPServer(AbstractChannel):
+    """A UDPServer is a communication channel. It allows to create a
+    server that listen to a specific IP:Port over a UDP socket.
 
-    When the actor execute an OpenChannelTransition, it calls the open
-    method on the tcp client which connects to the server.
+    When the actor executes an OpenChannelTransition, it calls the
+    open method on the UDP server which makes it to listen for
+    incomming messages.
 
     >>> from netzob.all import *
     >>> import time
-    >>> client = TCPClient(remoteIP='127.0.0.1', remotePort=9999)
+    >>> server = UDPServer(localIP='127.0.0.1', localPort=9999)
+    >>> server.open()
+    >>> server.close()
 
     >>> symbol = Symbol([Field("Hello Zoby !")])
     >>> s0 = State()
@@ -69,11 +71,11 @@ class TCPClient(AbstractChannel):
     >>> closeTransition = CloseChannelTransition(startState=s1, endState=s2)
     >>> automata = Automata(s0, [symbol])
 
-    >>> channel = TCPServer(localIP="127.0.0.1", localPort=8885)
+    >>> channel = UDPServer(localIP="127.0.0.1", localPort=8884)
     >>> abstractionLayer = AbstractionLayer(channel, [symbol])
     >>> server = Actor(automata = automata, initiator = False, abstractionLayer=abstractionLayer)
 
-    >>> channel = TCPClient(remoteIP="127.0.0.1", remotePort=8885)
+    >>> channel = UDPClient(remoteIP="127.0.0.1", remotePort=8884)
     >>> abstractionLayer = AbstractionLayer(channel, [symbol])
     >>> client = Actor(automata = automata, initiator = True, abstractionLayer=abstractionLayer)
 
@@ -86,32 +88,29 @@ class TCPClient(AbstractChannel):
 
     """
 
-    def __init__(self, remoteIP, remotePort, localIP=None, localPort=None, timeout=5):
-        super(TCPClient, self).__init__(isServer=False)
-        self.remoteIP = remoteIP
-        self.remotePort = remotePort
+    @typeCheck(str, int)
+    def __init__(self, localIP, localPort, timeout=5):
+        super(UDPServer, self).__init__(isServer=False)
         self.localIP = localIP
         self.localPort = localPort
         self.timeout = timeout
         self.__isOpen = False
         self.__socket = None
+        self.__remoteAddr = None
 
     def open(self, timeout=None):
-        """Open the communication channel. If the channel is a client, it starts to connect
-        to the specified server.
+        """Open the communication channel. This will open a UDP socket
+        that listen for incomming messages.
         """
 
         if self.isOpen:
-            raise RuntimeError("The channel is already open, cannot open it again")
+            raise RuntimeError("The channel is already open, cannot open it again.")
 
-        self.__socket = socket.socket()
+        self.__socket = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
         # Reuse the connection
         self.__socket.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
         self.__socket.settimeout(self.timeout)
-        if self.localIP is not None and self.localPort is not None:
-            self.__socket.bind((self.localIP, self.localPort))
-        self._logger.debug("Connect to the TCP server to {0}:{1}".format(self.remoteIP, self.remotePort))
-        self.__socket.connect((self.remoteIP, self.remotePort))
+        self.__socket.bind((self.localIP, self.localPort))
 
     def close(self):
         """Close the communication channel."""
@@ -119,46 +118,29 @@ class TCPClient(AbstractChannel):
             self.__socket.close()
 
     def read(self, timeout=None):
-        """Reads the next message on the communication channel.
-        Continues to read while it receives something.
-
+        """Read the next message on the communication channel.
 
         @keyword timeout: the maximum time in millisecond to wait before a message can be reached
         @type timeout: :class:`int`
         """
-        reading_seg_size = 1024
-        
+        # TODO: handle timeout
         if self.__socket is not None:
-            data = b""
-            finish = False
-            while not finish:
-                try:
-                    recv = self.__socket.recv(reading_seg_size)
-                except socket.timeout:
-                    # says we received nothing (timeout issue)
-                    recv = b""
-                if recv is None or len(recv) == 0:
-                    finish = True
-                else:
-                    data += recv
+            (data, self.__remoteAddr) = self.__socket.recvfrom(1024)
             return data
         else:
             raise Exception("socket is not available")
 
+    @typeCheck(bytes)
     def write(self, data):
         """Write on the communication channel the specified data
 
         :parameter data: the data to write on the channel
         :type data: binary object
         """
-        if self.__socket is not None:
-            try:
-                self.__socket.sendall(data)
-            except socket.error:
-                raise ChannelDownException()
-
+        if self.__socket is not None and self.__remoteAddr is not None:
+            self.__socket.sendto(data, self.__remoteAddr)
         else:
-            raise Exception("socket is not available")
+            raise Exception("Socket is not available or remote address is not known.")
 
     # Management methods
 
@@ -179,42 +161,6 @@ class TCPClient(AbstractChannel):
     # Properties
 
     @property
-    def remoteIP(self):
-        """IP on which the server will listen.
-
-        :type: :class:`str`
-        """
-        return self.__remoteIP
-
-    @remoteIP.setter
-    @typeCheck(str)
-    def remoteIP(self, remoteIP):
-        if remoteIP is None:
-            raise TypeError("RemoteIP cannot be None")
-
-        self.__remoteIP = remoteIP
-
-    @property
-    def remotePort(self):
-        """TCP Port on which the server will listen.
-        Its value must be above 0 and under 65535.
-
-
-        :type: :class:`int`
-        """
-        return self.__remotePort
-
-    @remotePort.setter
-    @typeCheck(int)
-    def remotePort(self, remotePort):
-        if remotePort is None:
-            raise TypeError("RemotePort cannot be None")
-        if remotePort <= 0 or remotePort > 65535:
-            raise ValueError("RemotePort must be > 0 and <= 65535")
-
-        self.__remotePort = remotePort
-
-    @property
     def localIP(self):
         """IP on which the server will listen.
 
@@ -225,12 +171,16 @@ class TCPClient(AbstractChannel):
     @localIP.setter
     @typeCheck(str)
     def localIP(self, localIP):
+        if localIP is None:
+            raise TypeError("ListeningIP cannot be None")
+
         self.__localIP = localIP
 
     @property
     def localPort(self):
-        """TCP Port on which the server will listen.
+        """UDP Port on which the server will listen.
         Its value must be above 0 and under 65535.
+
 
         :type: :class:`int`
         """
@@ -239,6 +189,11 @@ class TCPClient(AbstractChannel):
     @localPort.setter
     @typeCheck(int)
     def localPort(self, localPort):
+        if localPort is None:
+            raise TypeError("ListeningPort cannot be None")
+        if localPort <= 0 or localPort > 65535:
+            raise ValueError("ListeningPort must be > 0 and <= 65535")
+
         self.__localPort = localPort
 
     @property
