+++ src/netzob/Simulator/Channels/__init__.py
